Update style.css to fix opacity issue
There is a problem with the opacity fix (i.e. the previous pull request) in that it changes the background colors of the individual text pages (which is not desired). This problem is easily fixed through specifying that pages other than the text pages (which already have an id="other" attribute) are the only ones to get this opacity rgba background.